use plain text to save secrets to fix double auth failure.

+++ storage/model.go
@@ -3,7 +3,7 @@ package storage
 type (
 	Token struct {
 		AccessToken string `json:"accessToken"`
-		IssuedAt    int32  `json:"issuedAt"`
+		IssuedAt    int64  `json:"issuedAt"`
 	}
 
 	User struct {
@@ -11,8 +11,8 @@ type (
 		Name             string  `json:"name"`
 		Email            string  `json:"email"`
 		PushbulletID     string  `json:"indent"`
-		LastSeen         int32   `json:"lastSeen"`
-		CreatedAt        int32   `json:"created"`
+		LastSeen         int64   `json:"lastSeen"`
+		CreatedAt        int64   `json:"created"`
 		SimplePushSecret string  `json:"simplePushSecret"`
 		AdminSecret      string  `json:"adminSecret"`
 		Tokens           []Token `json:"tokens"`
